moved classes out to separate headers and added coroutine for the loop
